requirements.txt
# Requirements for Traffic Monitoring on Raspberry Pi
# Note: Install system packages first with install_system_packages.sh
# Then use: pip install -r requirements.txt

# Core packages that need pip installation
tqdm>=4.64.0

# Type checking packages (from original requirements)
mypy==1.0.1
mypy-extensions==0.4.3
pylint==2.16.2

# Development tools
astroid==2.14.2
isort==5.6.4
toml==0.10.2
tomlkit==0.11.7

# Web and networking
python-dotenv==1.2.1
cloud-sql-python-connector[pg8000]==1.18.5
pg8000==1.31.5
sqlalchemy==2.0.44
urllib3==1.26.12
certifi==2025.11.12
charset-normalizer==3.0.1
idna==3.3
ssl==1.16

# Additional utilities from original requirements
click==8.1.3
colorama==0.4.6
jedi==0.18.2
more-itertools==8.10.0
platformdirs==2.6.0
wrapt==1.14.1
zipp==1.0.0

# Data analysis and manipulation
seaborn==0.13.2
matplotlib==3.10.7
plotly==6.5.0
pandas==2.3.3
numpy==2.2.6
